backend/pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["Alina Mukhomediarova <alinoco.dia@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.6.0"
python-decouple = "^3.8"
requests = "^2.32.2"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.39"
django-extensions = "^3.2.3"
ipython = "^8.25.0"
django-model-utils = "^4.5.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
